[CIDMDEV-5110] - Removed persistence of Rackers from Identity
Implementation Notes
Rather than migrate the Daos and data for Rackers, the code will be modified to no longer require the persistence of Racker objects within Identity.

A pre-req for this is that all rackers will use AE tokens.

New Property
In order to accommodate  a mixed mode deployment, a new reloadable feature (feature.persist.rackers.enabled) is being added. When enabled, the code will persist rackers in the normal fashion so previous version nodes will be able to retrieve these rackers as normal. However, the new nodes will NOT rely on these persisted users in any way.

Feature Flag Removal
In addition to the above, the following racker related feature flags will be removed as the flags have been enabled in prod and were only added to facilitate mixed mode deployment :

feature.racker.username.auth.enabled : The code will be modified assuming this feature is always enabled.
feature.restrict.impersonation.to.rackers.with.role.enabled : The code will be modified assuming this feature is always enabled.

Docker Property Changes
The new property feature.persist.rackers.enabled has been set to 'true' by default in docker environments.

The existing properties feature.racker.defaultTokenFormat and feature.federated.defaultTokenFormat have been switched to 'AE' as it's a requirement that at least Rackers (including federated rackers) use AE tokens.

Deployment Notes
The flag feature.persist.rackers.enabled must initially be set to true on deployment. It must only be disabled when we are confident a rollback is not necessary. If a rollback does occur after racker persistence has been disabled, tokens issued to rackers who have never authenticated when persistence was enabled will not work (and result in a 404).

+++ src/test/groovy/com/rackspace/idm/api/resource/cloud/v20/FederatedRackerIntegrationTest.groovy
@@ -21,6 +21,7 @@ import org.openstack.docs.identity.api.v2.AuthenticateResponse
 import org.openstack.docs.identity.api.v2.BadRequestFault
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.test.context.ContextConfiguration
+import spock.lang.Unroll
 import testHelpers.IdmAssert
 import testHelpers.RootIntegrationTest
 import testHelpers.saml.SamlAssertionFactory
@@ -56,8 +57,11 @@ class FederatedRackerIntegrationTest extends RootIntegrationTest {
     def cleanup() {
     }
 
-    def "racker populated appropriately from saml and edir w/ no EDIR groups"() {
+    @Unroll
+    def "racker populated appropriately from saml and edir w/ no EDIR groups. Persist Racker: #persistRacker"() {
         given:
+        reloadableConfiguration.setProperty(IdentityConfig.FEATURE_PERSIST_RACKERS_PROP, persistRacker)
+
         def username = Constants.RACKER_NOGROUP
         def expDays = 5
 
@@ -71,20 +75,29 @@ class FederatedRackerIntegrationTest extends RootIntegrationTest {
         AuthenticateResponse authResponse = samlResponse.getEntity(AuthenticateResponse).value
         verifyResponseFromSamlRequest(authResponse, username)
 
-        when: "retrieve user from backend"
-        Racker fedUser = ldapFederatedRackerRepository.getUserByUsernameForIdentityProviderUri(authResponse.user.id, authResponse.user.federatedIdp)
-
-        then: "reflects current state"
-        fedUser.federatedUserName == authResponse.user.id
-        fedUser.federatedUserName == authResponse.user.name
-        fedUser.federatedIdpUri == RACKER_IDP_URI
+        and: "if persisted, the persisted object is valid"
+        if (persistRacker) {
+            Racker fedUser = ldapFederatedRackerRepository.getUserByUsernameForIdentityProviderUri(authResponse.user.id, authResponse.user.federatedIdp)
+            fedUser.federatedUserName == authResponse.user.id
+            fedUser.federatedUserName == authResponse.user.name
+            fedUser.federatedIdpUri == RACKER_IDP_URI
+        }
 
         cleanup:
-        deleteFederatedRackerQuietly(String.format("%s@%s", username, RACKER_IDP_URI))
+        if (persistRacker) {
+            deleteFederatedRackerQuietly(String.format("%s@%s", username, RACKER_IDP_URI))
+        }
+
+        where:
+        persistRacker | _
+        true | _
+        false | _
     }
 
-    def "racker populated appropriately from saml and edir w/ impersonate role"() {
+    @Unroll
+    def "racker populated appropriately from saml and edir w/ impersonate role. Persist Racker: #persistRacker"() {
         given:
+        reloadableConfiguration.setProperty(IdentityConfig.FEATURE_PERSIST_RACKERS_PROP, persistRacker)
         def username = Constants.RACKER_IMPERSONATE
         def expDays = 5
 
@@ -98,20 +111,29 @@ class FederatedRackerIntegrationTest extends RootIntegrationTest {
         AuthenticateResponse authResponse = samlResponse.getEntity(AuthenticateResponse).value
         verifyResponseFromSamlRequest(authResponse, username, [identityConfig.getStaticConfig().getRackerImpersonateRoleName()])
 
-        when: "retrieve user from backend"
-        Racker fedUser = ldapFederatedRackerRepository.getUserByUsernameForIdentityProviderUri(authResponse.user.id, authResponse.user.federatedIdp)
-
-        then: "reflects current state"
-        fedUser.federatedUserName == authResponse.user.id
-        fedUser.federatedUserName == authResponse.user.name
-        fedUser.federatedIdpUri == RACKER_IDP_URI
+        and: "if persisted, the persisted object is valid"
+        if (persistRacker) {
+            Racker fedUser = ldapFederatedRackerRepository.getUserByUsernameForIdentityProviderUri(authResponse.user.id, authResponse.user.federatedIdp)
+            fedUser.federatedUserName == authResponse.user.id
+            fedUser.federatedUserName == authResponse.user.name
+            fedUser.federatedIdpUri == RACKER_IDP_URI
+        }
 
         cleanup:
-        deleteFederatedRackerQuietly(String.format("%s@%s", username, RACKER_IDP_URI))
+        if (persistRacker) {
+            deleteFederatedRackerQuietly(String.format("%s@%s", username, RACKER_IDP_URI))
+        }
+
+        where:
+        persistRacker | _
+        true | _
+        false | _
     }
 
-    def "Validating token received matches initial federated auth response"() {
+    @Unroll
+    def "Validating token received matches initial federated auth response. Persist Racker: #persistRacker"() {
         given:
+        reloadableConfiguration.setProperty(IdentityConfig.FEATURE_PERSIST_RACKERS_PROP, persistRacker)
         def username = Constants.RACKER_IMPERSONATE
 
         def samlAssertion = new SamlAssertionFactory().generateSamlAssertionStringForFederatedRacker(RACKER_IDP_URI, username, 1);
@@ -142,10 +164,17 @@ class FederatedRackerIntegrationTest extends RootIntegrationTest {
 
         cleanup:
         deleteFederatedRackerQuietly(String.format("%s@%s", username, RACKER_IDP_URI))
+
+        where:
+        persistRacker | _
+        true | _
+        false | _
     }
 
-    def "Federated racker auth matches regular racker auth"() {
+    @Unroll
+    def "Federated racker auth matches regular racker auth. Persist Racker: #persistRacker"() {
         given:
+        reloadableConfiguration.setProperty(IdentityConfig.FEATURE_PERSIST_RACKERS_PROP, persistRacker)
         def username = Constants.RACKER_IMPERSONATE
         def samlAssertion = new SamlAssertionFactory().generateSamlAssertionStringForFederatedRacker(RACKER_IDP_URI, username, 1);
 
@@ -176,10 +205,17 @@ class FederatedRackerIntegrationTest extends RootIntegrationTest {
 
         cleanup:
         deleteFederatedRackerQuietly(String.format("%s@%s", username, RACKER_IDP_URI))
+
+        where:
+        persistRacker | _
+        true | _
+        false | _
     }
 
-    def "Federated racker w/ impersonation role can impersonate a user-admin and have token imp token validated"() {
+    @Unroll
+    def "Federated racker w/ impersonation role can impersonate a user-admin and have token imp token validated. Persist Racker: #persistRacker"() {
         given:
+        reloadableConfiguration.setProperty(IdentityConfig.FEATURE_PERSIST_RACKERS_PROP, persistRacker)
         def username = Constants.RACKER_IMPERSONATE
         def samlAssertion = new SamlAssertionFactory().generateSamlAssertionStringForFederatedRacker(RACKER_IDP_URI, username, 1);
 
@@ -205,6 +241,11 @@ class FederatedRackerIntegrationTest extends RootIntegrationTest {
 
         cleanup:
         deleteFederatedRackerQuietly(String.format("%s@%s", username, RACKER_IDP_URI))
+
+        where:
+        persistRacker | _
+        true | _
+        false | _
     }
 
     def "Invalid SAML signature results in 400"() {
